// types.ts
export interface TimelineMilestone {
  title: string;
  description: string;
  image: string;
  date?: string;
}

export interface UltrasoundData {
  src: string;
  date: string;
  weeks: string;
}

export interface LabelText {
  title: string;
  subtitle: string;
  description: string;
  vintage: string;
  variety: string;
  footerQuote: string;
}